+++ packages/bunshine/src/middleware/etags/etags.spec.ts
@@ -4,7 +4,6 @@ import HttpRouter from '../../HttpRouter/HttpRouter';
 import { etags } from './etags';
 
 describe('etags middleware', () => {
-  let port = 50800;
   const helloWorldEtag = '"c6a1dfa4103602ad"';
   let server: Server;
   let app: HttpRouter;